Fixed program - combined service classes and interface, refactored mode switcher etc
