--- src/main/java/io/codelex/flightplanner/admin/AdminService.java
@@ -1,11 +1,0 @@
-package io.codelex.flightplanner.admin;
-
-import io.codelex.flightplanner.flight.Flight;
-
-public interface AdminService {
-    Flight addFlight(Flight request);
-
-    Flight fetchFlight(long id);
-
-    void deleteFlight(long id);
-}
