+++ src/main/java/io/codelex/flightplanner/controller/AdminController.java
@@ -1,6 +1,8 @@
-package io.codelex.flightplanner.admin;
+package io.codelex.flightplanner.controller;
 
 import io.codelex.flightplanner.flight.*;
+import io.codelex.flightplanner.service.ServiceInterface;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -9,27 +11,29 @@ import org.springframework.web.bind.annotation.*;
 @RestController
 @RequestMapping("/admin-api/flights")
 public class AdminController {
-    private final AdminService adminService;
 
-    public AdminController(AdminService adminService) {
-        this.adminService = adminService;
+    @Autowired
+    private final ServiceInterface serviceClass;
+
+    public AdminController(ServiceInterface serviceClass) {
+        this.serviceClass = serviceClass;
     }
 
     @ResponseStatus(HttpStatus.CREATED)
     @PutMapping
     public Flight addFlight(@RequestBody Flight request) {
-        return adminService.addFlight(request);
+        return serviceClass.addFlight(request);
     }
 
     @ResponseStatus(HttpStatus.OK)
     @GetMapping("/{id}")
     public Flight fetchFlight(@PathVariable long id) {
-        return adminService.fetchFlight(id);
+        return serviceClass.fetchFlight(id);
     }
 
     @ResponseStatus(HttpStatus.OK)
     @DeleteMapping("/{id}")
     public void deleteFlight(@PathVariable long id) {
-        adminService.deleteFlight(id);
+        serviceClass.deleteFlight(id);
     }
 }
