+++ src/main/resources/application.properties
@@ -13,7 +13,7 @@ spring.jpa.database-platform=org.hibernate.dialect.H2Dialect
 
 spring.liquibase.change-log=classpath:/db/changelog/changelog-master.xml
 
-myapp.admin-storage-mode=database
-myapp.customer-storage-mode=database
+myapp.db-storage-mode=database
+myapp.in-mem-storage-mode=in-memory
 
 spring.jpa.show-sql=true
